# studio/CONTRACTS/openapi.yaml
openapi: 3.0.3
info:
  title: ClassmateFinder API
  version: 1.0.0
  description: API for ClassmateFinder - University student collaboration platform
  contact:
    name: ClassmateFinder Team
    email: tech@classmatefinder.ca

servers:
  - url: http://localhost:5000/api
    description: Development server

paths:
  /health:
    get:
      summary: Health check endpoint
      description: Returns API health status
      operationId: healthCheck
      tags: [Health]
      security: []
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time
                  environment:
                    type: string
                    example: development

  /auth/register:
    post:
      summary: Register a new user
      description: Create a new account with email verification required
      operationId: register
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/verify-email:
    post:
      summary: Verify email address
      description: Verify email using token sent to user's email
      operationId: verifyEmail
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyEmailResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      summary: Login with email and password
      description: Authenticate and receive JWT access token
      operationId: login
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/resend-verification:
    post:
      summary: Resend verification email
      description: Send a new verification token to the user's email
      operationId: resendVerification
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequest'
      responses:
        '200':
          description: Verification email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResendVerificationResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - username
        - password
        - firstName
      properties:
        email:
          type: string
          format: email
          example: user@ualberta.ca
        username:
          type: string
          minLength: 3
          maxLength: 30
          example: student123
        password:
          type: string
          minLength: 8
          example: SecurePass123!
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe

    RegisterResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
            username:
              type: string
            firstName:
              type: string
            lastName:
              type: string
            emailVerified:
              type: boolean
        requiresVerification:
          type: boolean
        message:
          type: string

    VerifyEmailRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          example: a1b2c3d4-e5f6-4a7b-8c9d-0e1f2a3b4c5d

    VerifyEmailResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        emailVerified:
          type: boolean

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@ualberta.ca
        password:
          type: string
          example: SecurePass123!

    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
            username:
              type: string
            emailVerified:
              type: boolean

    ResendVerificationRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: user@ualberta.ca

    ResendVerificationResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - INVALID_INPUT
            - INVALID_CREDENTIALS
            - EMAIL_NOT_VERIFIED
            - TOKEN_EXPIRED
            - TOKEN_INVALID
            - DOMAIN_NOT_ALLOWED
            - RATE_LIMITED
            - EMAIL_EXISTS
            - USER_NOT_FOUND
            - SERVER_ERROR
        message:
          type: string
        details:
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string

tags:
  - name: Health
    description: Health check endpoints
  - name: Authentication
    description: User authentication and authorization